omg that was such a nasty little bug. now i remember why i love tests. (the problem was that the employer restore was deleting old employer profiles, which often were the pre-seeded profile images. d'oh.

+++ employer/methods.py
@@ -83,7 +83,7 @@ class DataMethods():
 		for e in Employer.all().fetch(1000):
 			try:  
 				this_profile = Profile.get_by_key_name(e.unique_identifier)
-				this_profile.photo.delete()
+				if this_profile.photo.type != "pq": this_profile.photo.delete()
 			except: pass 	
 			
   def refresh_employer_images(self, employer_list=False):
